refactor: refactor message handling and update dependencies
- Update dependencies in `go.mod` to newer versions
- Add `github.com/jpillora/backoff` as an indirect dependency
- Change `queue.NewConsumer` to `queue.NewRing` in `main.go`
- Remove unnecessary type conversion for `WorkerNum` in `main.go`
- Add `Payload` method to `PushNotification` in `notification.go`
- Change `QueuedMessage` to `TaskMessage` in `SendNotification` function
- Modify JSON unmarshal to use `Payload` method in `SendNotification`
- Update `Run` function to use `TaskMessage` instead of `QueuedMessage`
- Update `server_test.go` to use `TaskMessage` instead of `QueuedMessage`
- Change data types in `status.go` for `BusyWorkers`, `SuccessTasks`, `FailureTasks`, and `SubmittedTasks`

+++ status/status.go
@@ -25,10 +25,10 @@ var StatStorage *StateStorage
 // App is status structure
 type App struct {
 	Version        string        `json:"version"`
-	BusyWorkers    int           `json:"busy_workers"`
-	SuccessTasks   int           `json:"success_tasks"`
-	FailureTasks   int           `json:"failure_tasks"`
-	SubmittedTasks int           `json:"submitted_tasks"`
+	BusyWorkers    int64         `json:"busy_workers"`
+	SuccessTasks   uint64        `json:"success_tasks"`
+	FailureTasks   uint64        `json:"failure_tasks"`
+	SubmittedTasks uint64        `json:"submitted_tasks"`
 	TotalCount     int64         `json:"total_count"`
 	Ios            IosStatus     `json:"ios"`
 	Android        AndroidStatus `json:"android"`
